+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to the "gpt-token-counter" extension will be documented in this file.
 
+## [1.2.4]
+
+### Changed
+- Minor fixes.
+
+## [1.2.3]
+
+### Added
+- Added support for new OpenAI models: o3-mini, o1, o1-mini, gpt-4o-mini
+
+### Changed
+- Updated to tiktoken 1.0.20
+- Updated Claude models to only include Claude-3.5, and Claude-3.7
+
+### Removed
+- Removed older models: text-davinci-003, davinci, babbage
+- Removed Claude-2 and Claude-3
+
 ## [1.2.2]
 
 ### Added